refactor: move sub menus to sub files
